// src/components/pages/dashboard/admin/customer-management/customer.type.ts
export type TCustomer = {
  _id: string;
  name: string;
  customerID: string;
  mobileNumber: string;
  email: string;
  companyName: string;
  user: string;
  gender: string;
  district: string;
  thana: string;
  homeAddress: string;
  salesExecutiveReference: string;
  createdAt: string;
  updatedAt: string;
  totalDue: number;
  totalPaid: number;
  totalPurchased: number;
};